Notification Archiving & Library Image Fix

+++ notifications.php
@@ -67,12 +67,16 @@ else {
 		print "</div>" ;
 	}
 
-	//Get notifications
+	print "<div class='linkTop'>" ; 
+		print "<a onclick='return confirm(\"Are you sure you want to delete these records.\")' title='" . _('Delete All Notifications') . "' href='" . $_SESSION[$guid]["absoluteURL"] . "/notificationsDeleteAllProcess.php'><img style='opacity: 0.8; vertical-align: -75%' src='" . $_SESSION[$guid]["absoluteURL"] . "/themes/" . $_SESSION[$guid]["gibbonThemeName"] . "/img/notifications_on_dark.png'></a>" ;
+	print "</div>" ;
+	
+	//Get and show newnotifications
 	try {
 		$dataNotifications=array("gibbonPersonID"=>$_SESSION[$guid]["gibbonPersonID"], "gibbonPersonID2"=>$_SESSION[$guid]["gibbonPersonID"]); 
-		$sqlNotifications="(SELECT gibbonNotification.*, gibbonModule.name AS source FROM gibbonNotification JOIN gibbonModule ON (gibbonNotification.gibbonModuleID=gibbonModule.gibbonModuleID) WHERE gibbonPersonID=:gibbonPersonID)
+		$sqlNotifications="(SELECT gibbonNotification.*, gibbonModule.name AS source FROM gibbonNotification JOIN gibbonModule ON (gibbonNotification.gibbonModuleID=gibbonModule.gibbonModuleID) WHERE gibbonPersonID=:gibbonPersonID AND status='New')
 		UNION
-		(SELECT gibbonNotification.*, 'System' AS source FROM gibbonNotification WHERE gibbonModuleID IS NULL AND gibbonPersonID=:gibbonPersonID2)
+		(SELECT gibbonNotification.*, 'System' AS source FROM gibbonNotification WHERE gibbonModuleID IS NULL AND gibbonPersonID=:gibbonPersonID2 AND status='New')
 		ORDER BY timestamp DESC, source, text" ;
 		$resultNotifications=$connection2->prepare($sqlNotifications);
 		$resultNotifications->execute($dataNotifications); 
@@ -82,26 +86,24 @@ else {
 	}
 
 	print "<h2>" ;
-		print _("Notifications") . " <span style='font-size: 65%; font-style: italic; font-weight: normal'> x" . $resultNotifications->rowCount() . "</span>" ;
+		print _("New Notifications") . " <span style='font-size: 65%; font-style: italic; font-weight: normal'> x" . $resultNotifications->rowCount() . "</span>" ;
 	print "</h2>" ;
-	print "<div class='linkTop'>" ; 
-		print "<a onclick='return confirm(\"Are you sure you want to delete these records.\")' title='" . _('Delete All Notifications') . "' href='" . $_SESSION[$guid]["absoluteURL"] . "/notificationsDeleteAllProcess.php'><img style='opacity: 0.8; vertical-align: -75%' src='" . $_SESSION[$guid]["absoluteURL"] . "/themes/" . $_SESSION[$guid]["gibbonThemeName"] . "/img/notifications_on_dark.png'>" ;
-	print "</div>" ;
+	
 	print "<table cellspacing='0' style='width: 100%'>" ;
 		print "<tr class='head'>" ;
-			print "<th>" ;
+			print "<th style='width: 18%'>" ;
 				print _("Source") ;
 			print "</th>" ;
-			print "<th>" ;
+			print "<th style='width: 12%'>" ;
 				print _("Date") ;
 			print "</th>" ;
-			print "<th>" ;
+			print "<th style='width: 51%'>" ;
 				print _("Message") ;
 			print "</th>" ;
-			print "<th>" ;
+			print "<th style='width: 7%'>" ;
 				print _("Count") ;
 			print "</th>" ;
-			print "<th style='width: 100px'>" ;
+			print "<th style='width: 12%'>" ;
 				print _("Actions") ;
 			print "</th>" ;
 		print "</tr>" ;
@@ -140,7 +142,85 @@ else {
 						print $row["count"] ;
 					print "</td>" ;
 					print "<td>" ;
-						print "<a href='" . $_SESSION[$guid]["absoluteURL"] . "/notificationsActionProcess.php?action=" . urlencode($row["actionLink"]) . "&gibbonNotificationID=" . $row["gibbonNotificationID"] . "'><img title='" . _('Action & Delete') . "' src='./themes/" . $_SESSION[$guid]["gibbonThemeName"] . "/img/plus.png'/></a> " ;
+						print "<a href='" . $_SESSION[$guid]["absoluteURL"] . "/notificationsActionProcess.php?action=" . urlencode($row["actionLink"]) . "&gibbonNotificationID=" . $row["gibbonNotificationID"] . "'><img title='" . _('Action & Archive') . "' src='./themes/" . $_SESSION[$guid]["gibbonThemeName"] . "/img/plus.png'/></a> " ;
+						print "<a href='" . $_SESSION[$guid]["absoluteURL"] . "/notificationsDeleteProcess.php?gibbonNotificationID=" . $row["gibbonNotificationID"] . "'><img title='" . _('Delete') . "' src='./themes/" . $_SESSION[$guid]["gibbonThemeName"] . "/img/garbage.png'/></a> " ;
+					print "</td>" ;
+				print "</tr>" ;
+			}
+		}
+	print "</table>" ;
+	
+	//Get and show newnotifications
+	try {
+		$dataNotifications=array("gibbonPersonID"=>$_SESSION[$guid]["gibbonPersonID"], "gibbonPersonID2"=>$_SESSION[$guid]["gibbonPersonID"]); 
+		$sqlNotifications="(SELECT gibbonNotification.*, gibbonModule.name AS source FROM gibbonNotification JOIN gibbonModule ON (gibbonNotification.gibbonModuleID=gibbonModule.gibbonModuleID) WHERE gibbonPersonID=:gibbonPersonID AND status='Archived')
+		UNION
+		(SELECT gibbonNotification.*, 'System' AS source FROM gibbonNotification WHERE gibbonModuleID IS NULL AND gibbonPersonID=:gibbonPersonID2 AND status='Archived')
+		ORDER BY timestamp DESC, source, text LIMIT 0, 50" ;
+		$resultNotifications=$connection2->prepare($sqlNotifications);
+		$resultNotifications->execute($dataNotifications); 
+	}
+	catch(PDOException $e) { 
+		print "<div class='error'>" . $e->getMessage() . "</div>" ; 
+	}
+
+	print "<h2>" ;
+		print _("Archived Notifications") ;
+	print "</h2>" ;
+	print "<table cellspacing='0' style='width: 100%'>" ;
+		print "<tr class='head'>" ;
+			print "<th style='width: 18%'>" ;
+				print _("Source") ;
+			print "</th>" ;
+			print "<th style='width: 12%'>" ;
+				print _("Date") ;
+			print "</th>" ;
+			print "<th style='width: 51%'>" ;
+				print _("Message") ;
+			print "</th>" ;
+			print "<th style='width: 7%'>" ;
+				print _("Count") ;
+			print "</th>" ;
+			print "<th style='width: 12%'>" ;
+				print _("Actions") ;
+			print "</th>" ;
+		print "</tr>" ;
+	
+		$count=0;
+		$rowNum="odd" ;
+		if ($resultNotifications->rowCount()<1) {
+			print "<tr class=$rowNum>" ;
+				print "<td colspan=5>" ;
+					print _("There are no records to display.") ;
+				print "</td>" ;
+			print "</tr>" ;
+		}
+		else {
+			while ($row=$resultNotifications->fetch() AND $count<20) {
+				if ($count%2==0) {
+					$rowNum="even" ;
+				}
+				else {
+					$rowNum="odd" ;
+				}
+				$count++ ;
+		
+				//COLOR ROW BY STATUS!
+				print "<tr class=$rowNum>" ;
+					print "<td>" ;
+						print $row["source"] ;
+					print "</td>" ;
+					print "<td>" ;
+						print dateConvertBack($guid, substr($row["timestamp"],0,10)) ;
+					print "</td>" ;
+					print "<td>" ;
+						print $row["text"] ;
+					print "</td>" ;
+					print "<td>" ;
+						print $row["count"] ;
+					print "</td>" ;
+					print "<td>" ;
+						print "<a href='" . $_SESSION[$guid]["absoluteURL"] . "/notificationsActionProcess.php?action=" . urlencode($row["actionLink"]) . "&gibbonNotificationID=" . $row["gibbonNotificationID"] . "'><img title='" . _('Action') . "' src='./themes/" . $_SESSION[$guid]["gibbonThemeName"] . "/img/plus.png'/></a> " ;
 						print "<a href='" . $_SESSION[$guid]["absoluteURL"] . "/notificationsDeleteProcess.php?gibbonNotificationID=" . $row["gibbonNotificationID"] . "'><img title='" . _('Delete') . "' src='./themes/" . $_SESSION[$guid]["gibbonThemeName"] . "/img/garbage.png'/></a> " ;
 					print "</td>" ;
 				print "</tr>" ;
